pyscripts/scripts/requirements.txt
absl-py==0.12.0
array-record==0.4.1
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
dm-tree==0.1.8
editdistance==0.6.2
emnist==0.0
etils==1.5.0
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.9.2
gast==0.5.4
google-api-core==2.12.0
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-cloud-vision==3.4.4
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio==1.59.0
grpcio-status==1.59.0
h5py==3.9.0
humanfriendly==10.0
idna==3.4
importlib-resources==6.1.0
joblib==1.3.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
matplotlib==3.8.0
ml-dtypes==0.2.0
mpmath==1.3.0
numpy==1.26.0
oauthlib==3.2.2
onnxruntime==1.16.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
path==16.7.1
Pillow==10.0.1
promise==2.3
proto-plus==1.22.3
protobuf==4.24.4
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-mnist==0.7
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-datasets==4.9.3
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.14.0
tensorflow-metadata==1.14.0
termcolor==2.3.0
tesseract==0.1.3
threadpoolctl==3.2.0
toml==0.10.2
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==3.0.0
wrapt==1.14.1
zipp==3.17.0
